+++ manifest/clients/ai21.py
@@ -82,6 +82,13 @@ class AI21Client(Client):
         """Return whether the client supports batch inference."""
         return False
 
+    def supports_streaming_inference(self) -> bool:
+        """Return whether the client supports streaming inference.
+
+        Override in child client class.
+        """
+        return False
+
     def get_model_params(self) -> Dict:
         """
         Get model params.