+++ manifest/clients/cohere.py
@@ -81,6 +81,13 @@ class CohereClient(Client):
         """Return whether the client supports batch inference."""
         return False
 
+    def supports_streaming_inference(self) -> bool:
+        """Return whether the client supports streaming inference.
+
+        Override in child client class.
+        """
+        return False
+
     def get_model_params(self) -> Dict:
         """
         Get model params.