+++ manifest/clients/dummy.py
@@ -48,6 +48,13 @@ class DummyClient(Client):
         """Return whether the client supports batch inference."""
         return True
 
+    def supports_streaming_inference(self) -> bool:
+        """Return whether the client supports streaming inference.
+
+        Override in child client class.
+        """
+        return False
+
     def get_generation_header(self) -> Dict[str, str]:
         """
         Get generation header.